Moved non navigation components out
