+++ src/components/TitleBar.tsx
@@ -2,12 +2,10 @@ import { StyleSheet, Text, View } from "react-native";
 import React from "react";
 import { styles } from "../styles";
 
-function TitleBar({ text }) {
+export const TitleBar = ({ text }) => {
   return (
     <View style={styles.titleBarBackground}>
       <Text style={styles.titleBar}>{text}</Text>
     </View>
   );
-}
-
-export { TitleBar };
+};
