+++ src/components/profile/NoProfile.tsx
@@ -1,6 +1,6 @@
 import { StyleSheet, Text, View } from "react-native";
 import React from "react";
-import { styles } from "../../../../styles";
+import { styles } from "../../styles";
 
 const NoProfile = () => (
   <View style={styles.container}>
