+++ src/navigation/tabs/Profile/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { View } from "react-native";
-import { Profile } from "./profile/Profile";
-import { NoProfile } from "./profile/NoProfile";
+import { Profile } from "../../../components/profile/Profile";
+import { NoProfile } from "../../../components/profile/NoProfile";
 import { useStateValue } from "../../../state";
 import { styles } from "../../../styles";
 import { TitleBar } from "../../../components/TitleBar";
