--- src/navigation/tabs/Settings/SettingsTab.tsx
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-
-import { View, Text } from "react-native";
-import { styles } from "../../../styles";
-
-export default class SettingsTab extends Component {
-  render() {
-    return (
-      <View style={styles.page}>
-        <Text>Settings Placeholder</Text>
-      </View>
-    );
-  }
-}
